pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "search-expert"
version = "0.1.0"
description = "Parse natural language search queries into structured fields using fine-tuned Qwen3.5-0.8B LoRA adapters."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Sarthak Rastogi" }]
requires-python = ">=3.10"
keywords = ["search", "nlp", "query-parsing", "llm", "structured-extraction", "fine-tuning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core runtime deps — model loading backends are optional extras
dependencies = [
    "torch>=2.0",
    "transformers>=5.5",
    "pyyaml>=6.0",
    "huggingface-hub>=0.22",
    "unsloth>=2026.4",
    
]

[project.optional-dependencies]
# Recommended: faster loading + 4-bit quant on GPU
unsloth = ["unsloth"]
# Alternative: standard HF PEFT without unsloth
peft = ["peft>=0.10", "bitsandbytes>=0.43"]
# Dev / test
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/sarthakrastogi/search-expert"
Repository = "https://github.com/sarthakrastogi/search-expert"
Issues = "https://github.com/sarthakrastogi/search-expert/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["search_expert*"]

# ── Ruff (linting + formatting) ───────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PIE", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["search_expert"]

# ── Pytest ────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
markers   = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
]

# ── Mypy ──────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict         = false
ignore_missing_imports = true